top_a return type is float

+++ exllamav2/exllamav2_ext/cpp/sampling.h
@@ -68,7 +68,7 @@ int top_a_cpu
     const int num_candidates,
     float* temp_probs,
     int* temp_indices,
-    int top_a
+    float top_a
 );
 
 int min_p_cpu
